+++ include/print_parameters.h
@@ -1,6 +1,6 @@
 #ifndef PRINT_PARAMETERS_H
 #define PRINT_PARAMETERS_H
 
-void print_parameters(int p, int q, int e, int d, int phi, int x, int m, int crypto_msg, int decrypto_msg);
+void print_parameters(long long p, long long q, long long e, long long d, long long phi, long long x, long long m, long long crypto_msg, long long decrypto_msg);
 
 #endif // PRINT_PARAMETERS_H
